We created CountScore method in Game class. This method counts the score of a player and sets it and it is used after every hand ends. We also tested it.

+++ Phase 10/Player.h
@@ -23,6 +23,6 @@ public:
 
 private:
 	std::string m_name;
-	uint8_t m_score;
+	uint16_t m_score;
 };
